requirements.txt

---

## Plik: requirements.txt

```text
torch
torchvision
matplotlib
Pillow
opencv-python
trimesh
open3d
